Extract a OrderCycleClone service
Before it was giving a Rubocop error, probably not a bad idea anyway.
