+++ app/models/order_cycle.rb
@@ -168,24 +168,7 @@ class OrderCycle < ApplicationRecord
   end
 
   def clone!
-    oc = dup
-    oc.name = I18n.t("models.order_cycle.cloned_order_cycle_name", order_cycle: oc.name)
-    oc.orders_open_at = oc.orders_close_at = oc.mails_sent = oc.processed_at = nil
-    oc.coordinator_fee_ids = coordinator_fee_ids
-    # rubocop:disable Layout/LineLength
-    oc.preferred_product_selection_from_coordinator_inventory_only = preferred_product_selection_from_coordinator_inventory_only
-    # rubocop:enable Layout/LineLength
-    oc.schedule_ids = schedule_ids
-    oc.save!
-    exchanges.each { |e| e.clone!(oc) }
-    oc.selected_distributor_payment_method_ids = (
-      attachable_distributor_payment_methods.map(&:id) & selected_distributor_payment_method_ids
-    )
-    oc.selected_distributor_shipping_method_ids = (
-      attachable_distributor_shipping_methods.map(&:id) & selected_distributor_shipping_method_ids
-    )
-    sync_subscriptions
-    oc.reload
+    OrderCycleClone.new(self).create
   end
 
   def variants
